Everything but the logger is ported
